+++ frontend/src/services/ant-design-pro/api.ts
@@ -32,14 +32,6 @@ export async function login(body: API.LoginParams, options?: { [key: string]: an
   });
 }
 
-/** 此处后端没有提供注释 GET /api/notices */
-export async function getNotices(options?: { [key: string]: any }) {
-  return request<API.NoticeIconList>('/api/notices', {
-    method: 'GET',
-    ...(options || {}),
-  });
-}
-
 /** 获取规则列表 GET /api/rule */
 export async function rule(
   params: {